- dmarc: install two more dependencies from ports - qmd: install p5-libwww first - haraka: set local_mx_ok in outbound.ini - host,base: improve cache emptying - i/mta: also disable sendmail periodic tasks - i/mua: move MUA tests into, used in dovecot & wildduck - mt: declare TLS_LIBRARY & TOASTER_BUILD_DEBUG - mt: TOASTER_PKG_AUDIT 0 -> 1 - provision/*: always declare JAIL_START_EXTRA, JAIL_CONF_EXTRA, JAIL_FSTAB - postfix: preserve main.cf (vs apply scripted settings every time) - roundcube: bump php version to 8.3 - spamassassin: remove NRPE plugin (port deleted) - squirrelmail: bump php 8.0 -> 8.1 - vpopmail: update source build - wildduck: add preflight_check - wildduck: add pf rules - zonemta: bump node version to 20

+++ provision/wildduck.sh
@@ -1,41 +1,88 @@
 #!/bin/sh
 
-. mail-toaster.sh || exit
+set -e
+
+. mail-toaster.sh
 
 export JAIL_START_EXTRA=""
 export JAIL_CONF_EXTRA=""
+export JAIL_FSTAB=""
 
+mt6-include mua
+
+preflight_check()
+{
+	for _j in dns redis mongodb
+	do
+		if ! jail_is_running "$_j"; then
+			fatal_err "jail $_j is required"
+		fi
+	done
+}
 
 install_webmail()
 {
-	stage_exec bash -c "cd /data && git clone https://github.com/nodemailer/wildduck-webmail.git webmail" || exit 1
-	stage_exec bash -c "cd /data/webmail && npm install"
-	stage_exec bash -c "cd /data/webmail && npm run bowerdeps"
+	if [ ! -e "$STAGE_MNT/data/webmail" ]; then
+		tell_status "installing wildduck webmail"
+		stage_exec bash -c "cd /data && git clone https://github.com/nodemailer/wildduck-webmail.git webmail" || exit 1
+		stage_exec bash -c "cd /data/webmail && npm install"
+		stage_exec bash -c "cd /data/webmail && npm run bowerdeps"
+	else
+		tell_status "updating wildduck webmail"
+		stage_exec bash -c "cd /data/webmail && git pull && npm install && npm run bowerdeps"
+		stage_exec bash -c "cd /data/webmail && mkdir -p public/components && bower install --allow-root"
+	fi
 }
 
 install_wildduck()
 {
-	tell_status "installing node.js"
-	stage_pkg_install npm-node16 git-tiny || exit
+	tell_status "installing wildduck dependencies"
+	stage_pkg_install npm-node20 git-tiny || exit
 
-	tell_status "installing wildduck"
-	stage_exec bash -c "cd /data && git clone https://github.com/nodemailer/wildduck.git" || exit 1
-	stage_exec bash -c "cd /data/wildduck && npm install --production"
+	if [ ! -e "$STAGE_MNT/data/webmail" ]; then
+		tell_status "installing wildduck"
+		stage_exec bash -c "cd /data && git clone https://github.com/nodemailer/wildduck.git" || exit 1
+		stage_exec bash -c "cd /data/wildduck && npm install --production"
+	else
+		tell_status "updating wildduck"
+		stage_exec bash -c "cd /data/wildduck && git pull && npm install --production"
+	fi
 
 	install_webmail
 }
 
+configure_pf()
+{
+	_pf_etc="$ZFS_DATA_MNT/wildduck/etc/pf.conf.d"
+
+	store_config "$_pf_etc/rdr.conf" <<EO_PF_RDR
+rdr proto tcp from any to <ext_ip4> port 993 -> $(get_jail_ip  wildduck) port 9993
+rdr proto tcp from any to <ext_ip4> port 995 -> $(get_jail_ip  wildduck) port 9995
+EO_PF_RDR
+
+	store_config "$_pf_etc/allow.conf" <<EO_PF_ALLOW
+mua_ports = "{ 993 995 9993 9995 }"
+table <mua_servers> persist { $(get_jail_ip wildduck), $(get_jail_ip6 wildduck) }
+pass in quick proto tcp from any to <mua_servers> port \$mua_ports
+EO_PF_ALLOW
+}
+
 configure_wildduck()
 {
-	sed -i '' \
-		-e "/^mongo/ s/127.0.0.1/$(get_jail_ip mongodb)/" \
-		-e "/^#redis/ s/127.0.0.1/$(get_jail_ip redis)/; s/\/3/\/8/" \
-		-e "/^host=/ s/127.0.0.1/$(get_jail_ip redis)/" \
-		-e "/^db=3/ s/3/8/" \
-		"$STAGE_MNT/data/wildduck/config/dbs.toml" || exit 1
+	_db_cfg="$STAGE_MNT/data/wildduck/config/dbs.toml"
+	if grep -qE '^mongo.*127' "$_db_cfg"; then
+		sed -i '' \
+			-e "/^mongo/ s/127.0.0.1/$(get_jail_ip mongodb)/" \
+			-e "/^#redis/ s/127.0.0.1/$(get_jail_ip redis)/; s/\/3/\/8/" \
+			-e "/^host=/ s/127.0.0.1/$(get_jail_ip redis)/" \
+			-e "/^db=3/ s/3/8/" \
+			"$_db_cfg"
+	fi
 
 	stage_exec npm install -g pm2
 	stage_exec pm2 startup
+
+	configure_pf
 }
 
 start_wildduck()
@@ -50,66 +97,24 @@ start_wildduck()
 	stage_exec pm2 save
 }
 
-test_imap()
-{
-	pkg install -y empty
-
-	POST_USER="postmaster@${TOASTER_MAIL_DOMAIN}"
-	POST_PASS=$(jexec vpopmail /usr/local/vpopmail/bin/vuserinfo -C "${POST_USER}")
-	rm -f in out
-
-	echo "testing IMAP AUTH as $POST_USER"
-
-	# empty -v -f -i in -o out telnet "$(get_jail_ip stage)" 143
-	empty -v -f -i in -o out openssl s_client -quiet -crlf -connect "$(get_jail_ip stage):9993"
-	if [ ! -e out ]; then exit; fi
-	empty -v -w -i out -o in "ready"             ". LOGIN $POST_USER $POST_PASS\n"
-	empty -v -w -i out -o in "Logged in"         ". LIST \"\" \"*\"\n"
-	empty -v -w -i out -o in "List completed"    ". SELECT INBOX\n"
-	# shellcheck disable=SC2050
-	if [ "has" = "some messages" ]; then
-		empty -v -w -i out -o in "Select completed"  ". FETCH 1 BODY\n"
-		empty -v -w -i out -o in "OK Fetch completed" ". LOGOUT\n"
-	else
-		empty -v -w -i out -o in "Select completed" ". LOGOUT\n"
-	fi
-	echo "Logout completed"
-	if [ -e out ]; then exit; fi
-}
-
-test_pop3()
-{
-	pkg install -y empty
-
-	POST_USER="postmaster@${TOASTER_MAIL_DOMAIN}"
-	POST_PASS=$(jexec vpopmail /usr/local/vpopmail/bin/vuserinfo -C "${POST_USER}")
-	rm -f in out
-
-	echo "testing POP3 AUTH as $POST_USER"
-
-	# empty -v -f -i in -o out telnet "$(get_jail_ip stage)" 110
-	empty -v -f -i in -o out openssl s_client -quiet -crlf -connect "$(get_jail_ip stage):9995"
-	if [ ! -e out ]; then exit; fi
-	empty -v -w -i out -o in "\+OK." "user $POST_USER\n"
-	empty -v -w -i out -o in "\+OK" "pass $POST_PASS\n"
-	empty -v -w -i out -o in "OK Logged in" "list\n"
-	empty -v -w -i out -o in "." "quit\n"
-
-	if [ -e out ]; then exit; fi
-}
-
 test_wildduck()
 {
 	tell_status "testing wildduck"
 	stage_listening 9993 3
 	stage_listening 9995 3
 	stage_listening 3000 3
+
+	# MUA_TEST_USER="postmaster@${TOASTER_MAIL_DOMAIN}"
+	# MUA_TEST_PASS=$(jexec vpopmail /usr/local/vpopmail/bin/vuserinfo -C "${MUA_TEST_USER}")
+	# MUA_TEST_HOST=$(get_jail_ip stage)
+
 	# test_imap
 	# test_pop3
 	echo "it worked"
 }
 
-base_snapshot_exists || exit
+base_snapshot_exists || exit 1
+preflight_check
 create_staged_fs wildduck
 start_staged_jail wildduck
 install_wildduck
